+++ py-polars/tests/unit/datatypes/test_temporal.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import contextlib
 import io
 from datetime import date, datetime, time, timedelta, timezone
-from typing import TYPE_CHECKING, Any, Literal, cast
+from typing import TYPE_CHECKING, Any, cast
 
 import numpy as np
 import pandas as pd
@@ -201,20 +201,6 @@ def test_from_pydatetime() -> None:
     assert s.name == "name"
     assert s.null_count() == 1
     assert s.dt[0] == dates[0]
-
-
-@pytest.mark.parametrize("time_unit", ["ms", "us", "ns"])
-def test_from_numpy_timedelta(time_unit: Literal["ns", "us", "ms"]) -> None:
-    s = pl.Series(
-        "name",
-        np.array(
-            [timedelta(days=1), timedelta(seconds=1)], dtype=f"timedelta64[{time_unit}]"
-        ),
-    )
-    assert s.dtype == pl.Duration(time_unit)
-    assert s.name == "name"
-    assert s.dt[0] == timedelta(days=1)
-    assert s.dt[1] == timedelta(seconds=1)
 
 
 def test_int_to_python_datetime() -> None:
